fix(core): 修复 Minecraft 1.21.10 兼容性问题
- 将依赖版本从 1.21.11 降级到 1.21.10
- 适配 FogRendererMixin 中 GpuBuffer.slice 方法参数类型变化
- 更新 RenderLayers 到 RenderLayer 并调整线条渲染逻辑
- 移除 VertexConsumer.lineWidth 调用，使用 RenderLayer.LINES 控制线宽
- 适配 MasterPlannerScreen.resize 方法签名变更

+++ src/main/java/com/masterplanner/mixin/FogRendererMixin.java
@@ -32,10 +32,11 @@ public class FogRendererMixin {
             // 返回空缓冲区，相当于禁用雾
             // 尝试使用不同的大小，从大到小，找到可用的最大值
             // slice 方法的 length 参数必须小于缓冲区大小（不能等于）
-            long[] sizes = {40L, 32L, 16L};
-            for (long size : sizes) {
+            // 1.21.10: GpuBuffer.slice 使用 int 参数
+            int[] sizes = {40, 32, 16};
+            for (int size : sizes) {
                 try {
-                    cir.setReturnValue(emptyBuffer.slice(0L, size));
+                    cir.setReturnValue(emptyBuffer.slice(0, size));
                     return;
                 } catch (IllegalArgumentException e) {
                     // 继续尝试下一个更小的值
